Automate publishing to GitHub Pages
I am following along this tutorial[1] in order to set up my repository
so that I only have to keep track of the actual org files that I am
editing. Even though I probably won't push to this site very often, it
is in my best interest to not waste energy building over and over again.

[1]: https://systemcrafters.net/publishing-websites-with-org-mode/automated-site-publishing/