Set the publishing-directory to "./public"
In order for the site to build, we must build it in the public
directory. Changing this value allows us to have a clear distinction
between the host and me building the site. Originally I had used the
docs/ folder so that I could use that with the 'Build and deployment'
settings in GitHub, but now that I continue to follow this[1] tutorial I
am changing it.

[1]: https://systemcrafters.net/publishing-websites-with-org-mode/automated-site-publishing/